Revert "Merge pull request #65 from oss-slu/custom_roles"
This reverts commit c6d36e236b51379ef5de031d0774f8b5d7627f2d, reversing
changes made to 097bd11024e0a44208ee4b1b010a568e45a7f713.
