+++ components/classroom/fetchData.mjs
@@ -1,4 +1,4 @@
-import TPEN from '../../api/TPEN.mjs';
+import TPEN from '../../TPEN/index.mjs';
 
 const PROJECT_FORM = document.getElementById("projectId");
 const MSG_CONTAINER = document.getElementById("msg"); 
@@ -52,3 +52,8 @@ document.getElementById("clickList").addEventListener("click", function (event) 
 });
 
 fetchProjectData(PROJECT_FORM);
+
+
+
+
+
